Improved irma-shelters.json add/remove detection
As of https://github.com/Irma-Response/irma-api/pull/23 an id property is
exposed in the https://irma-api.herokuapp.com/api/v1/shelters API.

I'm now using that to perform more accurate detection of added/removed
shelters. Previously I woud treat a shelter as "removed and then added" if the
name of the shelter changed. Now I can detect that as an update to an existing
shelter instead.