Fixed logic bug in de-duplication
We were ignoring any duplicate groups that included at least one shelter that
had been marked as not-a-dupe.

This meant that if a new duplicate shelter was added it would be ignored.

Instead, we now filter out "marked as not dupe" shelters before we build the
potential duplicate lists, fixing this bug.