Detect florida-shelters missing from irma-shelters
This code compares the shelters listed in florida-shelters.json with the
shelters we have manually added to irma-shelters.json via the
https://irma-api.herokuapp.com/ tool. It generates a list of potentially
missing shelters, based on geohash location comparisons.

These can be fixed either by adding the missing shelter, or by pasting the map
URL of the suggested shelter into a comment on https://github.com/simonw/irma-
scraped-data/issues/2 to mark it as whitelisted.

Changes to the list of missing shelters will be published to Slack.

References https://github.com/simonw/irma-scraped-data/issues/1